Renaming the text to speech file
